扩展体力值模块，根据用户登录时间差动态计算用户当前体力值，并整合后台管理端和websocket，实现健康状态推送和管理员推送 注意nginx需要设置，serverEndpointExpotser在外置tomcat下要注释

+++ src/main/java/com/mt/metro/annotation/CacheLock.java
@@ -20,7 +20,7 @@ public @interface CacheLock {
     /**
      * 过期秒数,默认为5秒
      */
-    int expire() default 5;
+    long expire() default 5;
 
     /**
      * 超时时间单位，默认为秒
